MAINT: MockOptimizer moved to optmizers.py, such that it can be reused for tests
